# Server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for backing up videos.",
        "title": "WeatherReels",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v2",
    "paths": {
        "/api/v2/tags/create": {
            "post": {
                "description": "Gets all of the tags associated with a video",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos",
                    "tags"
                ],
                "summary": "Get Video's Tags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video id",
                        "name": "request",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tag"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/videos/abort": {
            "post": {
                "description": "Gets a group of video and thumbnail urls with the given offset. Max of 10",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a group of video/thumbnail urls",
                "parameters": [
                    {
                        "description": "Video information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetVideoGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetVideoGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/videos/description": {
            "put": {
                "description": "Updates a given videoID's description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Updates a videos description",
                "parameters": [
                    {
                        "description": "New Title",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangeVideoDescriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/videos/fetch/{video}": {
            "get": {
                "description": "Generate a presigned URL for a video file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get URL for a video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video information",
                        "name": "request",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/videos/random": {
            "get": {
                "description": "Gets all the required data for a random video display page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get random data for a video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video information",
                        "name": "request",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetRandomVideoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/videos/title": {
            "put": {
                "description": "Updates a given videoID's title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Updates a videos title",
                "parameters": [
                    {
                        "description": "New Title",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangeVideoTitleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/videos/verify": {
            "put": {
                "description": "Verifies videos that are already uploaded, returning a list of ones that are not verified",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Verify uploaded Videos",
                "parameters": [
                    {
                        "description": "List of video filenames",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/videos/video": {
            "get": {
                "description": "Gets all the required data for a video display page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get data for a video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Video information",
                        "name": "request",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetVideoDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ChangeVideoDescriptionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "row_id": {
                    "type": "string"
                }
            }
        },
        "models.ChangeVideoTitleRequest": {
            "type": "object",
            "properties": {
                "row_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.GetRandomVideoResponse": {
            "type": "object",
            "properties": {
                "row_id": {
                    "type": "string"
                }
            }
        },
        "models.GetVideoDataResponse": {
            "type": "object",
            "properties": {
                "custom_description": {
                    "type": "string"
                },
                "custom_title": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "row_id": {
                    "type": "string"
                },
                "s3_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
        "models.GetVideoGroupCursor": {
            "type": "object",
            "properties": {
                "row_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.GetVideoGroupPartResponse": {
            "type": "object",
            "properties": {
                "custom_description": {
                    "type": "string"
                },
                "custom_title": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "row_id": {
                    "type": "string"
                },
                "s3_id": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.GetVideoGroupRequest": {
            "type": "object",
            "properties": {
                "row_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.GetVideoGroupResponse": {
            "type": "object",
            "properties": {
                "cursor": {
                    "$ref": "#/definitions/models.GetVideoGroupCursor"
                },
                "videos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetVideoGroupPartResponse"
                    }
                }
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.Tag": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}